Operator: fix crash on invalid policy.
If operator got a invalid policy there was a crash, with this change
operator will log the error.

Sample policy
```
apiVersion: cilium.io/v2
kind: CiliumNetworkPolicy
metadata:
  name: allow-egress-ingress-cluster
spec:
  egress:
  - toEndpoints:
    - matchLabels:
        'k8s:io.kubernetes.pod.namespace': {}
  endpointSelector:
    matchLabels: {}
  ingress:
  - fromEndpoints:
    - matchLabels:
        'k8s:io.kubernetes.pod.namespace': {}
```

Evidence:
```
NAMESPACE     NAME                                       READY   STATUS    RESTARTS   AGE
kube-system   pod/cilium-etcd-8hmkpwk8xs                 1/1     Running   0          36m
kube-system   pod/cilium-etcd-l8h9rlspc5                 1/1     Running   0          35m
kube-system   pod/cilium-etcd-lkh4wfvj8z                 1/1     Running   0          37m
kube-system   pod/cilium-etcd-operator-d6cb86868-pg8wf   1/1     Running   0          37m
kube-system   pod/cilium-operator-cb9fb7b5-lqnvn         1/1     Running   1          22m
kube-system   pod/cilium-tqdrf                           1/1     Running   1          37m
kube-system   pod/coredns-fff89c9b9-cb8dt                1/1     Running   0          45m
kube-system   pod/etcd-k8s1                              1/1     Running   0          44m
kube-system   pod/etcd-operator-5cf67779fd-r7kjm         1/1     Running   0          37m
kube-system   pod/kube-apiserver-k8s1                    1/1     Running   0          44m
kube-system   pod/kube-controller-manager-k8s1           1/1     Running   0          44m
kube-system   pod/kube-proxy-ljr7q                       1/1     Running   0          45m
kube-system   pod/kube-scheduler-k8s1                    1/1     Running   0          44m

NAMESPACE   NAME                                                         AGE
default     ciliumnetworkpolicy.cilium.io/allow-egress-ingress-cluster   21m
```

Fix #6910
