+++ daemon/k8s_watcher_test.go
@@ -18,7 +18,6 @@ package main
 
 import (
 	"net"
-	"time"
 
 	"github.com/cilium/cilium/pkg/annotation"
 	"github.com/cilium/cilium/pkg/checker"
@@ -43,27 +42,6 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/tools/cache"
 )
-
-func (ds *DaemonSuite) TestK8sErrorLogTimeout(c *C) {
-	errstr := "I am an error string"
-
-	// ensure k8sErrMsg is empty for tests that use it
-	k8sErrMsgMU.Lock()
-	k8sErrMsg = map[string]time.Time{}
-	k8sErrMsgMU.Unlock()
-
-	// Returns true because it's the first time we see this message
-	startTime := time.Now()
-	c.Assert(k8sErrorUpdateCheckUnmuteTime(errstr, startTime), Equals, true)
-
-	// Returns false because <= k8sErrLogTimeout time has passed
-	noLogTime := startTime.Add(k8sErrLogTimeout)
-	c.Assert(k8sErrorUpdateCheckUnmuteTime(errstr, noLogTime), Equals, false)
-
-	// Returns true because k8sErrLogTimeout has passed
-	shouldLogTime := startTime.Add(k8sErrLogTimeout).Add(time.Nanosecond)
-	c.Assert(k8sErrorUpdateCheckUnmuteTime(errstr, shouldLogTime), Equals, true)
-}
 
 func (ds *DaemonSuite) Test_missingK8sNetworkPolicyV1(c *C) {
 	type args struct {
